// frontend/src/app/api/model/earthquake-list-response.ts
/**
 * Earthquake Monitor API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EarthquakeResponse } from './earthquake-response';


/**
 * Schema for paginated earthquake list responses
 */
export interface EarthquakeListResponse { 
    /**
     * List of earthquake records
     */
    earthquakes: Array<EarthquakeResponse>;
    /**
     * Total number of records matching the filter
     */
    total: number;
    /**
     * Current page number
     */
    page: number;
    /**
     * Number of records per page
     */
    limit: number;
    /**
     * Whether there are more pages available
     */
    has_next: boolean;
    /**
     * Whether there are previous pages available
     */
    has_previous: boolean;
}