// frontend/src/app/api/model/earthquake-filter.ts
/**
 * Earthquake Monitor API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Schema for filtering earthquake records
 */
export interface EarthquakeFilter { 
    min_magnitude?: number | null;
    max_magnitude?: number | null;
    min_depth?: number | null;
    max_depth?: number | null;
    start_time?: string | null;
    end_time?: string | null;
    place_contains?: string | null;
    magnitude_type?: string | null;
}