// frontend/src/app/api/model/pagination-params.ts
/**
 * Earthquake Monitor API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Schema for pagination parameters
 */
export interface PaginationParams { 
    /**
     * Page number (starts at 1)
     */
    page?: number;
    /**
     * Number of records per page (max 1000)
     */
    limit?: number;
}